feat: #6/파일명 변경 - PropertyReferenceException - Entity명과 동일한 Repository이어야 한다는 Exception으로 인해 파일명 변경
